Added bin executables, made a refactor and added some documentation
