+++ bin/list-devices
@@ -1,0 +1,2 @@
+#!/usr/bin/env node
+require('../list.js');
