+++ bin/sms
@@ -1,0 +1,2 @@
+#!/usr/bin/env node
+require('../index.js')(process.argv[2], process.argv[3]);
