+++ package.json
@@ -1,22 +1,13 @@
 {
   "name": "rpi-sms",
-<<<<<<< HEAD
   "version": "0.0.1",
-  "description": "Enviando SMS com RaspberryPI + Modem3G.",
-  "main": "app.js",
+  "description": "Send SMS messages with RaspberryPI + 3G Modem.",
+  "main": "index.js",
   "dependencies": {
-    "serialport": "~1.3.1"
-=======
-  "version": "0.0.0",
-  "main": "app.js",
-  "dependencies": {
-    "serialport": "^1.3.1"
->>>>>>> 83e32e13be2e286d5ab177bbe3cd11b6281be68c
-  },
+    "serialport": "~1.3.1"  },
   "devDependencies": {},
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
-<<<<<<< HEAD
   },
   "repository": {
     "type": "git",
@@ -28,7 +19,4 @@
     "url": "https://github.com/rogerin/rpi-sms/issues"
   },
   "homepage": "https://github.com/rogerin/rpi-sms"
-=======
-  }
->>>>>>> 83e32e13be2e286d5ab177bbe3cd11b6281be68c
 }
